Move ops/tree/format/parse to util/format
