+++ src/cargo/util/format.rs
